// src/types/MultiStepWizzardTypes.ts
type MultiStepFormDataTypes = {
  uploadedFiles?: File[];
  imagePreviews?: string[];
  studioName?: string;
  studioAddress?: {
    street?: string;
    city?: string;
    zip?: string;
    country?: string;
  };
  contactNumber?: string;
  studioSize?: number;
  studioCapacity?: number;
  studioDescription?: string;
  amenities?: string[];
  studioAvailability?: {
    startDate?: string;
    endDate?: string;
    weeklySchedule?: { day: string; from: string; to: string }[];
  };
  // Motion Expert Fields
  sessionTypes?: string[];
  sessionTitles?: string[];
  sessionDescriptions?: string[];
  sessionTags?: string[];
  sessionDuration?: number;
  sessionPrice?: number;
  sessionPricing?: number;
  locationType?: "studio-host" | "self-host";
  selfHostedLocationDetails?: {
    locationName?: string;
    locationAddress?: {
      street?: string;
      city?: string;
      zip?: string;
      country?: string;
    };
    locationDescription?: string;
    indoor: boolean;
    outdoor: boolean;
    maxCapacity?: number;
    AGBAccepted?: boolean;
  };
};
export default MultiStepFormDataTypes;